# hub/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kanban Board API",
    "description": "Kanban board backend API with Better Auth and Socket.IO",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/UserWithPermissions"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "operationId": "listUsersMe"
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/UserWithPermissions"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "updateUsersMe"
      }
    },
    "/api/users/me/password": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Change password (Better Auth)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Validation or policy error"
          },
          "401": {
            "description": "Current password incorrect"
          },
          "501": {
            "description": "Password change not available"
          }
        },
        "operationId": "createUsersMePassword"
      }
    },
    "/api/users/me/preferences": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get notification/locale preferences",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Preferences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "preferences": {
                      "$ref": "#/components/schemas/UserPreferences"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "listUsersMePreferences"
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update preferences (partial)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferencesPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated preferences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "preferences": {
                      "$ref": "#/components/schemas/UserPreferences"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "updateUsersMePreferences"
      }
    },
    "/api/users/me/settings-layout": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get persisted Settings Hub card layouts",
        "description": "Returns stored JSON or null if never saved. Aligns with Kanban-frontend PersistedSettingsLayoutsV1.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Layout document or none",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "layout": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/SettingsHubLayoutsV1"
                        }
                      ],
                      "nullable": true,
                      "type": "object",
                      "description": "Settings layout document (null if never saved)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "listUsersMeSettingslayout"
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Replace settings hub layouts",
        "description": "Full document replace. version must be 1; userId must match session user. Unknown page or card ids return 400.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsLayoutPutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved layout (lastUpdatedAt set server-side)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "layout": {
                      "$ref": "#/components/schemas/SettingsHubLayoutsV1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid layout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "updateUsersMeSettingslayout"
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Clear server-stored layouts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "deleteUsersMeSettingslayout"
      }
    },
    "/api/users/me/sessions": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List active sessions for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions (non-expired, newest first, max 50)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserSessionSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "getSession1"
      }
    },
    "/api/users/me/sessions/revoke-others": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Revoke all other sessions (keep current Bearer token)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Other sessions revoked"
          },
          "401": {
            "description": "Unauthorized or missing Bearer token"
          }
        },
        "operationId": "createUsersMeSessionsRevokeothers"
      }
    },
    "/api/users/me/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get one session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session": {
                      "$ref": "#/components/schemas/UserSessionDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid sessionId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "operationId": "getSession2"
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Revoke a session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoked"
          },
          "400": {
            "description": "Invalid sessionId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "operationId": "deleteUsersMeSessions"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token received from login/register endpoints"
      },
      "agentAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-agent-api-key",
        "description": "Agent API key authentication using the x-agent-api-key header. Provide the raw key value (for example: ag_xxx). Agents must have active delegations to access projects."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or category"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "SUPPORT",
          "ADMIN"
        ],
        "description": "User's system role: USER (default), SUPPORT (elevated), ADMIN (full access)"
      },
      "UserPermissions": {
        "type": "object",
        "description": "User's system-level permissions based on their role",
        "properties": {
          "isAdmin": {
            "type": "boolean",
            "description": "Whether the user has admin privileges"
          },
          "canManageRateLimits": {
            "type": "boolean",
            "description": "Whether the user can manage rate limit configurations"
          },
          "canManageUsers": {
            "type": "boolean",
            "description": "Whether the user can manage user roles"
          },
          "canManageSystem": {
            "type": "boolean",
            "description": "Whether the user has full system access"
          }
        }
      },
      "UserWithPermissions": {
        "type": "object",
        "description": "User object with role and permissions for frontend access control",
        "properties": {
          "id": {
            "type": "integer",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User's display name"
          },
          "fullName": {
            "type": "string",
            "description": "User's full name (alias for name)"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "permissions": {
            "$ref": "#/components/schemas/UserPermissions"
          }
        }
      },
      "UserEmailNotifications": {
        "type": "object",
        "properties": {
          "taskAssigned": {
            "type": "boolean"
          },
          "taskCommented": {
            "type": "boolean"
          },
          "dailyDigest": {
            "type": "boolean"
          }
        }
      },
      "UserPreferences": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "minLength": 2,
            "maxLength": 10
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "emailNotifications": {
            "$ref": "#/components/schemas/UserEmailNotifications"
          }
        }
      },
      "UserPreferencesPatch": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "minLength": 2,
            "maxLength": 10
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "emailNotifications": {
            "type": "object",
            "properties": {
              "taskAssigned": {
                "type": "boolean"
              },
              "taskCommented": {
                "type": "boolean"
              },
              "dailyDigest": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "UserSessionSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Session id as string in JSON"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "isCurrent": {
            "type": "boolean"
          }
        }
      },
      "UserSessionDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserSessionSummary"
          },
          {
            "type": "object",
            "properties": {
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "isExpired": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
      "SettingsLayoutCardPlacement": {
        "type": "object",
        "required": [
          "id",
          "x",
          "y",
          "w",
          "h"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Known card id e.g. account.profile"
          },
          "x": {
            "type": "integer",
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "minimum": 0
          },
          "w": {
            "type": "integer",
            "minimum": 1,
            "maximum": 24
          },
          "h": {
            "type": "integer",
            "minimum": 1,
            "maximum": 80
          },
          "hidden": {
            "type": "boolean"
          }
        }
      },
      "SettingsLayoutPage": {
        "type": "object",
        "required": [
          "grid",
          "cards"
        ],
        "properties": {
          "grid": {
            "type": "object",
            "required": [
              "columns"
            ],
            "properties": {
              "columns": {
                "type": "integer",
                "minimum": 1,
                "maximum": 24
              }
            }
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettingsLayoutCardPlacement"
            }
          }
        }
      },
      "SettingsHubLayoutsV1": {
        "type": "object",
        "required": [
          "version",
          "userId",
          "lastUpdatedAt",
          "pages"
        ],
        "properties": {
          "version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "userId": {
            "type": "string",
            "description": "Must match authenticated user id"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pages": {
            "type": "object",
            "properties": {
              "account": {
                "$ref": "#/components/schemas/SettingsLayoutPage"
              },
              "agents": {
                "$ref": "#/components/schemas/SettingsLayoutPage"
              },
              "project": {
                "$ref": "#/components/schemas/SettingsLayoutPage"
              },
              "admin": {
                "$ref": "#/components/schemas/SettingsLayoutPage"
              },
              "themeBuilder": {
                "$ref": "#/components/schemas/SettingsLayoutPage"
              }
            },
            "additionalProperties": false
          }
        }
      },
      "SettingsLayoutPutRequest": {
        "type": "object",
        "required": [
          "layout"
        ],
        "properties": {
          "layout": {
            "$ref": "#/components/schemas/SettingsHubLayoutsV1"
          }
        }
      },
...
    }
  }
}